requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
googlemaps==4.10.0
pandas==2.1.4
numpy==1.26.2
firebase-admin==6.5.0

